// src/types.ts
export interface UseViewModelInstanceEnumResult {
  value: string | null;
  values: string[];
  setValue: (value: string) => void;
}